requirements.txt
Flask
PyYAML
langdetect
scikit-learn
rapidfuzz
flashtext
requests
geopy
